requirements.txt
streamlit
pandas
python-docx
PyPDF2
scikit-learn
xgboost
matplotlib
seaborn
